+++ src/main/java/dto/PlayerResultDto.java
@@ -2,7 +2,7 @@ package dto;
 
 public record PlayerResultDto(
     String name,
-    double profit
+    int profit
 ) {
 
 }
